[WIN-47][#4]🚑Asset 위치 WineyKit으로 수정
