+++ Projects/App/Sources/WineyApp.swift
@@ -1,11 +1,11 @@
 import SwiftUI
-import Utils
+import WineyKit
 
 @main
 struct WineyApp: App {
   
   init() {
-    UtilsFontFamily.registerAllCustomFonts() //프로젝트 전체에서 공유할 수 있도록
+    WineyKitFontFamily.registerAllCustomFonts() //프로젝트 전체에서 공유할 수 있도록
   }
   
   var body: some Scene {
