+++ Projects/Utils/Project.swift
@@ -1,17 +1,6 @@
 import ProjectDescription
 import ProjectDescriptionHelpers
 
-let infoPlist: [String: InfoPlist.Value] = [
-  "CFBundleShortVersionString": "1.0",
-  "CFBundleVersion": "1",
-  "CFBundleDevelopmentRegion": "ko_KR",
-  "UIUserInterfaceStyle": "Dark",
-  "UIAppFonts": [
-    "Item 0": "Pretendard-Medium.otf",
-    "Item 1": "Pretendard-Bold.otf"
-  ]
-]
-
 
 let utilsTargets: [Target] = [
     .init(name: "Utils",
@@ -19,9 +8,7 @@ let utilsTargets: [Target] = [
           product: .framework,
           bundleId: "com.adultOfNineteen.utils",
           deploymentTarget: .iOS(targetVersion: "15.0", devices: [.iphone]),
-          infoPlist: .extendingDefault(with: infoPlist),
           sources: ["Sources/**"],
-          resources: ["Resources/**"],
           dependencies: [
           ]
          )
